cleaned up route code
